Breaking change: Changed and centralized server configuration process
 - Both server and standalone client use blobs::Configuration to configure logging and server parameters
 - Removed blobs::InitializeServerLogging()
 - blobs::Initialize() now accepts a blobs::Configuration instead of the databsae root dir
 - This will make it easier to add more configuration values in the future withoutadding lots of parameters everywhere and/or breaking the API again

+++ Network/common/Encoding.cpp
@@ -8,17 +8,17 @@ namespace blobs::encoding {
 
 std::wstring ToUTF16(std::string_view content) {
   std::wstring result;
-  auto requiredSize = MultiByteToWideChar(CP_UTF8, MB_PRECOMPOSED, content.data(), content.size(), nullptr, 0);
+  auto requiredSize = MultiByteToWideChar(CP_UTF8, MB_PRECOMPOSED, content.data(), static_cast<int>(content.size()), nullptr, 0);
   result.resize(requiredSize);
-  MultiByteToWideChar(CP_UTF8, MB_PRECOMPOSED, content.data(), content.size(), result.data(), result.size());
+  MultiByteToWideChar(CP_UTF8, MB_PRECOMPOSED, content.data(), static_cast<int>(content.size()), result.data(), static_cast<int>(result.size()));
   return result;
 }
 
 std::string ToUTF8(std::wstring_view content) {
   std::string result;
-  auto requiredSize = WideCharToMultiByte(CP_UTF8, NULL, content.data(), content.size(), nullptr, 0, nullptr, nullptr);
+  auto requiredSize = WideCharToMultiByte(CP_UTF8, NULL, content.data(), static_cast<int>(content.size()), nullptr, 0, nullptr, nullptr);
   result.resize(requiredSize);
-  WideCharToMultiByte(CP_UTF8, NULL, content.data(), content.size(), result.data(), result.size(), nullptr, nullptr);
+  WideCharToMultiByte(CP_UTF8, NULL, content.data(), static_cast<int>(content.size()), result.data(), static_cast<int>(result.size()), nullptr, nullptr);
   return result;
 }
